requirements.txt
opencv-python~=4.11.0.86
tensorflow~=2.9.0
pillow~=11.1.0
keras~=2.9.0
scikit-learn
matplotlib
numpy~=1.26.4
flask~=3.1.0
torch==1.13.1
torchvision~=0.14.1
captcha==0.11